Fixed a bunch of issues
Fixed most (all?) of the problems with the redguard ship meshes. The shipwrecks are in terrible shape and have not been gone over yet. A number of the textures lack proper HD variants and do not tile as well.
Added proper collision meshes to the Imperial runestones.
Lengthened the Shroud of Namira's 1st person mesh so that it can not be seen through in some animations.
Fixed the extra nodes in the headless skeleton mesh and removed head from its shadow mesh.
Improved geometry of wenbone pauldron and removed stencil + changed alpha properties for the feathers.
Made the hautoma icon slightly darker, since it brighter than the item.