Add flag 32 to *most* switchnode meshes
On `NiSwitchNode` and its subclasses, such as `NiLODNode`, the flag 32, `UpdateOnlyActive`, tells Morrowind.exe not to spend any CPU time updating stuff like UV animations for branches of the switchnode which are not currently visible. It turns out that this makes a significant performance difference for scenes with a lot of hidden animations, such as Glow in the Dahrk-aware meshes during the daytime. For instance, in the worst spot in Old Ebonheart, this pushed FPS from about 60 to about 65–67.

Greatness7 provided a script to automatically add flag 32 to most of our switchnodes. The script skipped over any meshes with a `NiParticleSystemController` inside, because it seems that not updating particlesystems can in some cases cause them to look weird when they become visible. The particle-containing switchnode meshes will have to be tested and edited separately.